refactor: Preview 파일 분리 제거
