+++ Dialog/core/designsystem/src/commonMain/kotlin/com/on/dialog/designsystem/theme/Shadow.kt
@@ -1,6 +1,20 @@
 package com.on.dialog.designsystem.theme
 
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.dropShadow
 import androidx.compose.ui.graphics.Color
@@ -10,7 +24,6 @@ import androidx.compose.ui.graphics.shadow.Shadow
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
-import com.on.dialog.designsystem.theme.preview.ShadowPreview
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 enum class ShadowLevel(
@@ -44,6 +57,53 @@ fun Modifier.dropShadow(
 @Composable
 private fun ShadowPreview() {
     DialogTheme {
-        ShadowPreview()
+        Surface(modifier = Modifier.fillMaxSize()) {
+            Column(
+                modifier = Modifier
+                    .verticalScroll(rememberScrollState())
+                    .padding(24.dp),
+                verticalArrangement = Arrangement.spacedBy(32.dp),
+            ) {
+                Text(
+                    text = "Shadow Tokens",
+                    style = MaterialTheme.typography.headlineMedium,
+                    modifier = Modifier.padding(bottom = 8.dp),
+                )
+
+                ShadowItem("NONE", ShadowLevel.NONE)
+                ShadowItem("SMALL", ShadowLevel.SMALL)
+                ShadowItem("MEDIUM", ShadowLevel.MEDIUM)
+                ShadowItem("LARGE", ShadowLevel.LARGE)
+                ShadowItem("EXTRA_LARGE", ShadowLevel.EXTRA_LARGE)
+            }
+        }
+    }
+}
+
+@Composable
+private fun ShadowItem(
+    name: String,
+    level: ShadowLevel,
+) {
+    Column(
+        modifier = Modifier.fillMaxWidth(),
+        verticalArrangement = Arrangement.spacedBy(8.dp),
+    ) {
+        Text(
+            text = name,
+            style = MaterialTheme.typography.bodyMedium,
+        )
+        Box(
+            modifier = Modifier
+                .size(80.dp)
+                .dropShadow(level = level, shape = MaterialTheme.shapes.medium)
+                .background(MaterialTheme.colorScheme.surface),
+            contentAlignment = Alignment.Center,
+        ) {
+            Text(
+                text = name,
+                style = MaterialTheme.typography.labelSmall,
+            )
+        }
     }
 }
