+++ Dialog/core/designsystem/src/commonMain/kotlin/com/on/dialog/designsystem/theme/Shape.kt
@@ -1,10 +1,25 @@
 package com.on.dialog.designsystem.theme
 
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Shapes
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.unit.dp
-import com.on.dialog.designsystem.theme.preview.ShapePreview
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 internal val DialogShapes =
@@ -20,6 +35,56 @@ internal val DialogShapes =
 @Composable
 private fun ShapePreview() {
     DialogTheme {
-        ShapePreview()
+        Surface(modifier = Modifier.fillMaxSize()) {
+            Column(
+                modifier = Modifier
+                    .verticalScroll(rememberScrollState())
+                    .padding(16.dp),
+                verticalArrangement = Arrangement.spacedBy(16.dp),
+            ) {
+                Text(
+                    text = "Shape Tokens",
+                    style = MaterialTheme.typography.headlineMedium,
+                    modifier = Modifier.padding(bottom = 8.dp),
+                )
+
+                ShapeItem("extraSmall", MaterialTheme.shapes.extraSmall)
+                ShapeItem("small", MaterialTheme.shapes.small)
+                ShapeItem("medium", MaterialTheme.shapes.medium)
+                ShapeItem("large", MaterialTheme.shapes.large)
+                ShapeItem("extraLarge", MaterialTheme.shapes.extraLarge)
+            }
+        }
+    }
+}
+
+@Composable
+private fun ShapeItem(
+    name: String,
+    shape: Shape,
+) {
+    Column(
+        modifier = Modifier.fillMaxWidth(),
+        verticalArrangement = Arrangement.spacedBy(8.dp),
+    ) {
+        Text(
+            text = name,
+            style = MaterialTheme.typography.bodyMedium,
+        )
+        Box(
+            modifier = Modifier
+                .size(100.dp, 60.dp)
+                .background(
+                    color = MaterialTheme.colorScheme.primaryContainer,
+                    shape = shape,
+                ),
+            contentAlignment = Alignment.Center,
+        ) {
+            Text(
+                text = name,
+                style = MaterialTheme.typography.labelSmall,
+                color = MaterialTheme.colorScheme.onPrimaryContainer,
+            )
+        }
     }
 }
